Refactor AssuranceOptions class and move it to upgrade.py
